- fixed fseek offset signedness problem on x86_64
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code/trunk/gwyddion@5698 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ libgwymodule/gwymodule-file.c
@@ -421,7 +421,7 @@ gwy_file_detect_fill_info(GwyFileDetectInfo *fileinfo,
     fileinfo->head = g_new0(guchar, 2*fileinfo->buffer_len);
     fileinfo->tail = fileinfo->head + fileinfo->buffer_len;
     if (fread((gchar*)fileinfo->head, fileinfo->buffer_len - 1, 1, fh) < 1
-        || fseek(fh, 1-fileinfo->buffer_len, SEEK_END) != 0
+        || fseek(fh, 1-(gint)fileinfo->buffer_len, SEEK_END) != 0
         || fread((gchar*)fileinfo->tail, fileinfo->buffer_len - 1, 1, fh) < 1) {
         fclose(fh);
         gwy_file_detect_free_info(fileinfo);
